python-face-service/requirements.txt
# Python 3.10+ recommended. First run downloads model weights (large download).
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
deepface>=0.0.93
opencv-python-headless>=4.10.0
Pillow>=10.0.0
numpy>=1.26.0
